Add Publisher metadata to Book
