+++ ReadingList/Models/ModelVersions/books.xcdatamodeld/books_14.xcdatamodel/contents
@@ -14,6 +14,7 @@
         <attribute name="notes" optional="YES" attributeType="String" syncable="YES"/>
         <attribute name="pageCount" optional="YES" attributeType="Integer 32" minValueString="0" usesScalarValueType="YES" syncable="YES"/>
         <attribute name="publicationDate" optional="YES" attributeType="Date" usesScalarValueType="NO" syncable="YES"/>
+        <attribute name="publisher" optional="YES" attributeType="String" syncable="YES"/>
         <attribute name="rating" optional="YES" attributeType="Integer 16" minValueString="1" maxValueString="5" usesScalarValueType="YES" syncable="YES"/>
         <attribute name="readState" attributeType="Integer 16" defaultValueString="2" usesScalarValueType="YES" syncable="YES"/>
         <attribute name="sort" optional="YES" attributeType="Integer 32" usesScalarValueType="YES" syncable="YES"/>
@@ -71,7 +72,7 @@
         </fetchIndex>
     </entity>
     <elements>
-        <element name="Book" positionX="-63" positionY="-18" width="128" height="345"/>
+        <element name="Book" positionX="-63" positionY="-18" width="128" height="360"/>
         <element name="List" positionX="-54" positionY="117" width="128" height="90"/>
         <element name="Subject" positionX="-63" positionY="90" width="128" height="75"/>
     </elements>
